% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codes.R
\name{diag.RBS}
\alias{diag.RBS}
\alias{diag.BP}
\alias{residuals.pearson}
\alias{res_pearson}
\title{Diagnostic Analysis - Local Influnce}
\usage{
diag.BP(model, mu.link = "log", sigma.link = "log", scheme = "case.weight")

res_pearson(model)
}
\arguments{
\item{model}{Object of class \code{gamlss} holding the fitted model.}

\item{mu.link}{Defines the mu.link, with "identity" link as the default for the mu parameter.}

\item{sigma.link}{Defines the sigma.link, with "identity" link as the default for the sigma parameter}

\item{scheme}{Default is "case.weight". But, can be "response", "location" or "precision".}
}
\value{
Local influence measures.
}
\description{
Diagnostics for the BP model
}
\references{
Leiva, V., Santos-Neto, M., Cysneiros, F.J.A, Barros, M. (2014)  Birnbaum-Saunders statistical modelling: a new approach. \emph{Statistical Modelling}, v. 14, p. 21-48, 2014.
}
\author{
Manoel Santos-Neto \email{manoel.ferreira@ufcg.edu.br}, F.J.A. Cysneiros \email{cysneiros@de.ufpe.br}, Victor Leiva \email{victorleivasanchez@gmail.com} and Michelli Barros \email{michelli.karinne@gmail.com}
}
