% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\alias{grad_test_bp}
\alias{grad_test_ga}
\alias{grad_test_ig}
\alias{grad_test_wei}
\alias{wald_test}
\alias{score_test_bp}
\alias{score_test_ga}
\alias{score_test_ig}
\alias{score_test_wei}
\title{Precision test}
\usage{
grad_test_bp(modelh0,modelh1)

grad_test_ga(modelh0, modelh1)

grad_test_ig(modelh0, modelh1)

grad_test_wei(modelh0, modelh1)

score_test_bp(modelh0, modelh1)

score_test_ga(modelh0, modelh1)

score_test_ig(modelh0, modelh1)

score_test_wei(modelh0, modelh1)

wald_test(modelh1)
}
\arguments{
\item{modelh0}{model under null hypothesis.}

\item{modelh1}{model under alternative hypothesis.}
}
\value{
A list with class "htest" containing the following components:

\code{statistic}	the value of the test statistic.

\code{parameter}	the degrees of freedom for the test statistic.

\code{p.value}	the p-value for the test.

\code{method}	a character string indicating what type of likelihood
ratio test was performed.

\code{data.name} a character string giving the name(s) of the data.
}
\description{
Tests the null hypothesis of precision fixed in GA, IG, WEI and BP regression models against the alternative of precision variable.
}
\author{
Manoel Santos-Neto \email{manoel.ferreira@professor.ufcg.edu.br}
}
