% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Codes.R
\name{BP}
\alias{BP}
\alias{dBP}
\alias{pBP}
\alias{qBP}
\alias{rBP}
\alias{hBP}
\alias{plotBP}
\alias{meanBP}
\title{Reparameterized Beta Prime (BP) distribution for fitting a GAMLSS}
\usage{
BP(mu.link = "log", sigma.link = "log")

dBP(x, mu = 1, sigma = 1, log = FALSE)

pBP(q, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)

rBP(n, mu = 1, sigma = 1)

qBP(p, mu = 1, sigma = 1, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{mu.link}{object for which the extraction of model residuals is meaningful.}

\item{sigma.link}{type of residual to be used.}

\item{x, q}{vector of quantiles}

\item{mu}{vector of scale parameter values}

\item{sigma}{vector of shape parameter values}

\item{lower.tail}{logical; if TRUE (default), probabilities are P[X <= x], otherwise, P[X > x]}

\item{log.p, }{log.p logical; if TRUE, probabilities p are given as log(p).}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken to be the number required.}

\item{p}{vector of probabilities.}

\item{log.d, }{log.p logical; if TRUE, density d are given as log(d).}

\item{log.q, }{log.p logical; if TRUE, quantiles q are given as log(q).}

\item{log.h, }{log.p logical; if TRUE, hazard h are given as log(h).}

\item{from}{where to start plotting the distribution from.}

\item{to}{up to where to plot the distribution.}

\item{obj}{a fitted RWL object.}

\item{...}{other graphical parameters for plotting.}

\item{title}{title of the plot.}
}
\value{
returns a \code{gamlss.family} object which can be used to fit a reparameterized weighted Lindley distribution in the \code{gamlss()} function.
}
\description{
The fuction \code{BP()} defines the BP distribution, a two paramenter
distribution, for a gamlss.family object to be used in GAMLSS fitting using using the
function \code{gamlss()}, with mean equal to the parameter \code{mu} and \code{sigma}
equal the precision parameter. The functions \code{dBP}, \code{pBP}, \code{qBP},
\code{rBP} and \code{hBP} define the density, distribution function, quantile function, random
genetation and hazard function for the \code{BP} parameterization of the RWL distribution.
}
\note{
For the function RWL(), mu is the mean and sigma is the precision parameter of the reparameterized weighted Lindley distribution.
}
\author{
Manoel Santos-Neto \email{santosnetoce at protemail.com}
}
